refactor: change package structure.
